Software/requirements.txt
pywebview>=4.0
Pillow>=10.0
requests>=2.28
qrcode>=7.4
